src/server/virtual.museum.adminui/scripts/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
uuid==1.30
